+++ src/noshow/api/app_helpers.py
@@ -1,6 +1,12 @@
 import pickle
 from pathlib import Path
-from typing import Any, Union
+from typing import Any, List, Union
+
+from pandas import Series
+from sqlalchemy import select
+from sqlalchemy.orm import Session
+
+from noshow.database.models import ApiPrediction
 
 
 def load_model(model_path: Union[str, Path, None] = None) -> Any:
@@ -30,3 +36,34 @@ def add_clinic_phone(clinic_name: str) -> str:
         return "53594"
     else:
         return ""
+
+
+def fix_outdated_appointments(
+    session: Session, app_ids: Union[List, Series], start_date: str
+) -> None:
+    """Set the status of outdated appointments on inactive
+
+    Appointments can change while the model is running, existing apointments
+    will be updated by the api, but deleted appointments or appointments that
+    are rescheduled further in the future need to be set to inactive to prevent
+    them from showing up in the dashboard.
+
+    Parameters
+    ----------
+    session : Session
+        Session variable that holds the database connection
+    app_ids : Union[List, Series]
+        List of app ids for which a prediction has been made
+    start_date : str
+        start date from which the predictions are made
+    """
+    all_ids = session.execute(
+        select(ApiPrediction.id).where(ApiPrediction.start_time > start_date)
+    ).all()
+    inactive_ids = set(all_ids).difference(app_ids)
+    for app_id in inactive_ids:
+        apiprediction = session.get(ApiPrediction, app_id)
+        if apiprediction:
+            apiprediction.active = False
+            session.merge(apiprediction)
+            session.commit()